pyproject.toml
[tool.poetry]
name = "my_hackathon_bot"
version = "0.1.0"
description = ""
authors = ["WebDev23"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
python-telegram-bot = "^13.15"

[tool.poetry.group.dev.dependencies]
ipython = "^8.17.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
